# Simplify DCEL label management