--- a/geom/dcel_extract_intersection_matrix.go
+++ b/geom/dcel_extract_intersection_matrix.go
@@ -21,29 +21,25 @@ func (d *doublyConnectedEdgeList) extractIntersectionMatrix() matrix {
 }
 
 func (f *faceRecord) location(operand operand) imLocation {
-	assertPresence(f.labels)
-	if !f.labels[operand].inSet {
+	if !f.inSet[operand] {
 		return imExterior
 	}
 	return imInterior
 }
 
 func (e *halfEdgeRecord) location(operand operand) imLocation {
-	assertPresence(e.edgeLabels)
-
-	if !e.edgeLabels[operand].inSet {
-		return imExterior
+	face1Present := e.incident.inSet[operand]
+	face2Present := e.twin.incident.inSet[operand]
+	if face1Present && face2Present {
+		return imInterior
 	}
-
-	assertPresence(e.incident.labels)
-	assertPresence(e.twin.incident.labels)
-
-	face1Present := e.incident.labels[operand].inSet
-	face2Present := e.twin.incident.labels[operand].inSet
 	if face1Present != face2Present {
 		return imBoundary
 	}
-	return imInterior
+	if e.inSet[operand] {
+		return imInterior
+	}
+	return imExterior
 }
 
 func (v *vertexRecord) location(operand operand) imLocation {