--- a/geom/dcel_label.go
+++ b/geom/dcel_label.go
@@ -1,7 +1,5 @@
 package geom
 
-import "fmt"
-
 // operand represents either the first (A) or second (B) geometry in a binary
 // operation (such as Union or Covers).
 type operand int
@@ -11,40 +9,14 @@ const (
 	operandB operand = 1
 )
 
-type label struct {
-	// Set to true once inSet has a valid value.
-	populated bool
-
-	// Indicates whether the thing being labelled is in a set or not (context
-	// specific to how the label is used).
-	inSet bool
+func forEachOperand(fn func(operand operand)) {
+	fn(operandA)
+	fn(operandB)
 }
 
-func newUnpopulatedLabels() [2]label {
-	return [2]label{}
-}
-
-func newHalfPopulatedLabels(operand operand, inSet bool) [2]label {
-	var labels [2]label
-	labels[operand].populated = true
-	labels[operand].inSet = inSet
-	return labels
-}
-
-func newPopulatedLabels(inSet bool) [2]label {
-	var labels [2]label
-	labels[0].populated = true
-	labels[1].populated = true
-	labels[0].inSet = inSet
-	labels[1].inSet = inSet
-	return labels
-}
-
-func mergeLabels(dst *[2]label, src [2]label) {
-	dst[0].populated = dst[0].populated || src[0].populated
-	dst[1].populated = dst[1].populated || src[1].populated
-	dst[0].inSet = dst[0].inSet || src[0].inSet
-	dst[1].inSet = dst[1].inSet || src[1].inSet
+func mergeBools(dst *[2]bool, src [2]bool) {
+	dst[0] = dst[0] || src[0]
+	dst[1] = dst[1] || src[1]
 }
 
 type location struct {
@@ -63,30 +35,4 @@ func newLocationsOnBoundary(operand operand) [2]location {
 	var locs [2]location
 	locs[operand].boundary = true
 	return locs
-}
-
-func assertPresence(labels [2]label) {
-	if !labels[0].populated || !labels[1].populated {
-		panic(fmt.Sprintf("all presence flags in labels not set: %v", labels))
-	}
-}
-
-func selectUnion(labels [2]label) bool {
-	assertPresence(labels)
-	return labels[0].inSet || labels[1].inSet
-}
-
-func selectIntersection(labels [2]label) bool {
-	assertPresence(labels)
-	return labels[0].inSet && labels[1].inSet
-}
-
-func selectDifference(labels [2]label) bool {
-	assertPresence(labels)
-	return labels[0].inSet && !labels[1].inSet
-}
-
-func selectSymmetricDifference(labels [2]label) bool {
-	assertPresence(labels)
-	return labels[0].inSet != labels[1].inSet
 }