--- a/geom/dcel_overlay.go
+++ b/geom/dcel_overlay.go
@@ -30,14 +30,15 @@ func (d *doublyConnectedEdgeList) overlay(other *doublyConnectedEdgeList) {
 	d.overlayEdges(other)
 	d.fixVertices()
 	d.reAssignFaces()
-	d.fixLabels()
+	d.populateInSetLabels()
 }
 
 func (d *doublyConnectedEdgeList) overlayVertices(other *doublyConnectedEdgeList) {
 	for xy, otherVert := range other.vertices {
 		vert, ok := d.vertices[xy]
 		if ok {
-			mergeLabels(&vert.labels, otherVert.labels)
+			mergeBools(&vert.src, otherVert.src)
+			mergeBools(&vert.inSet, otherVert.inSet)
 			mergeLocations(&vert.locations, otherVert.locations)
 		} else {
 			d.vertices[xy] = otherVert
@@ -69,8 +70,9 @@ func (d *doublyConnectedEdgeList) overlayEdges(other *doublyConnectedEdgeList) {
 
 	for _, e := range other.halfEdges {
 		if existing, ok := edges.lookupEdge(e); ok {
-			mergeLabels(&existing.edgeLabels, e.edgeLabels)
-			mergeLabels(&existing.faceLabels, e.faceLabels)
+			mergeBools(&existing.srcEdge, e.srcEdge)
+			mergeBools(&existing.srcFace, e.srcFace)
+			mergeBools(&existing.inSet, e.inSet)
 		} else {
 			edges.insertEdge(e)
 			e.origin = d.vertices[e.origin.coords]
@@ -129,67 +131,52 @@ func (d *doublyConnectedEdgeList) reAssignFaces() {
 	d.faces = nil
 	for _, cycle := range cycles {
 		f := &faceRecord{
-			cycle:  cycle,
-			labels: [2]label{}, // populated below
+			cycle: cycle,
 		}
 		d.faces = append(d.faces, f)
 		forEachEdge(cycle, func(e *halfEdgeRecord) {
-			mergeLabels(&f.labels, e.faceLabels)
+			forEachOperand(func(operand operand) {
+				if e.srcFace[operand] {
+					f.inSet[operand] = true
+				}
+			})
 			e.incident = f
 		})
 	}
+
+	// Populate inSet for faces that did not have edges from their respective
+	// input geometries.
+	forEachOperand(func(operand operand) {
+		visited := make(map[*faceRecord]bool)
+		var dfs func(*faceRecord)
+		dfs = func(f *faceRecord) {
+			if visited[f] {
+				return
+			}
+			visited[f] = true
+			forEachEdge(f.cycle, func(e *halfEdgeRecord) {
+				if !e.srcFace[operand] {
+					e.twin.incident.inSet[operand] = true
+					dfs(e.twin.incident)
+				}
+			})
+		}
+		for _, f := range d.faces {
+			if f.inSet[operand] {
+				dfs(f)
+			}
+		}
+	})
 
 	// If we couldn't find any cycles, then we wouldn't have constructed any
 	// faces. This happens in the case where there are only point geometries.
 	// We need to artificially create an infinite face.
 	if len(d.faces) == 0 {
 		d.faces = append(d.faces, &faceRecord{
-			cycle:  nil,
-			labels: newPopulatedLabels(false),
+			cycle: nil,
+			inSet: [2]bool{},
 		})
 	}
-
-	for _, face := range d.faces {
-		d.completePartialFaceLabel(face)
-	}
-}
-
-// completePartialFaceLabel checks to see if the face label for the given face
-// is complete (i.e. contains a part for both A and B). If it's not complete,
-// then in searches adjacent faces until it finds a face that it can copy the
-// missing part of the label from. This situation occurs whenever a face in the
-// overlay DCEL doesn't have any edges from one of the original geometries.
-func (d *doublyConnectedEdgeList) completePartialFaceLabel(face *faceRecord) {
-	labelIsComplete := func() bool {
-		return face.labels[0].populated && face.labels[1].populated
-	}
-	if labelIsComplete() {
-		return
-	}
-	expanded := make(map[*faceRecord]bool)
-	stack := []*faceRecord{face}
-	for len(stack) > 0 {
-		popped := stack[len(stack)-1]
-		stack = stack[:len(stack)-1]
-		adjacent := adjacentFaces(popped)
-		expanded[popped] = true
-		for _, adj := range adjacent {
-			face.labels = completeLabels(face.labels, adj.labels)
-			if labelIsComplete() {
-				return
-			}
-			if !expanded[adj] {
-				stack = append(stack, adj)
-			}
-		}
-	}
-
-	// It's possible that we're still missing part of the face label. This
-	// could happen if one of the inputs is a Point/MultiPoint input because
-	// its associated ghost lines would not add to the label pool. We can
-	// safely fill in the presence bits for this case.
-	face.labels[0].populated = true
-	face.labels[1].populated = true
 }
 
 // adjacentFaces finds all of the faces that adjacent to f.
@@ -206,36 +193,18 @@ func adjacentFaces(f *faceRecord) []*faceRecord {
 	return adjacent
 }
 
-// completeLabels copies any missing portion (part A or part B) of the label
-// from donor to recipient, and then returns recipient.
-func completeLabels(recipient, donor [2]label) [2]label {
-	for i := 0; i < 2; i++ {
-		if !recipient[i].populated && donor[i].populated {
-			recipient[i].populated = true
-			recipient[i].inSet = donor[i].inSet
-		}
-	}
-	return recipient
-}
-
-// fixLabels updates edge and vertex labels after performing an overlay.
-func (d *doublyConnectedEdgeList) fixLabels() {
+// populateInSetLabels populates the inSet labels for edges and vertices.
+func (d *doublyConnectedEdgeList) populateInSetLabels() {
 	for _, e := range d.halfEdges {
 		// Copy labels from incident faces into edge since the edge represents
 		// the (closed) border of the face.
-		mergeLabels(&e.edgeLabels, e.incident.labels)
-		mergeLabels(&e.edgeLabels, e.twin.incident.labels)
-
-		// If we haven't seen an edge label yet for one of the two input
-		// geometries, we can assume that we'll never see it. So we mark off
-		// that side as having the bit populated.
-		e.edgeLabels[0].populated = true
-		e.edgeLabels[1].populated = true
+		e.inSet[0] = e.srcEdge[0] || e.incident.inSet[0] || e.twin.incident.inSet[0]
+		e.inSet[1] = e.srcEdge[1] || e.incident.inSet[1] || e.twin.incident.inSet[1]
 
 		// Copy edge labels onto the labels of adjacent vertices. This is
 		// because the vertices represent the endpoints of the edges, and
 		// should have at least those bits set.
-		mergeLabels(&e.origin.labels, e.edgeLabels)
-		mergeLabels(&e.origin.labels, e.prev.edgeLabels)
+		e.origin.inSet[0] = e.origin.src[0] || e.inSet[0] || e.prev.inSet[0]
+		e.origin.inSet[1] = e.origin.src[1] || e.inSet[1] || e.prev.inSet[1]
 	}
 }