## Description

commit 45ccc4bdfc05aa867cdddaa2061dd8da79c4de83
Author: Peter Stace <peterstace@gmail.com>
Date:   Mon Nov 7 20:19:20 2022 +1100

    Simplify DCEL label management
    
    In the DCEL data structure, there are labels on the vertex records,
    (half) edge records, and face records. These labels keep track
    information such as which input geometry the record originates from and
    whether it will appear in the output geometry.
    
    The original approach to labelling was to use the same data structure
    for all record types. The label had two booleans per operand.  The first
    boolean determined whether the record would appear in the final output,
    and the second boolean would show whether the first boolean was
    populated or not. This was essentially tri-state logic. The trouble with
    this approach was that the DCEL algorithm doesn't know whether each
    record should appear in the output geometry right up until the last
    stage. The labels are used in the earlier stages of the algorithm to
    record other facts, such as whether or not the records appear in the
    input geometry. The semantics of the fields subtly change over the
    course of the algorithm, making the code hard to understand and
    maintain.
    
    The new approach (implemented by this PR) changes how label management
    is performed. Instead of having a single labels data structure, the
    fields are slightly different for vertex, edge, and face records. This
    is to account for the differences in information available when
    processing vertexes vs. edges vs. faces. The number of fields is also
    expanded, so that concepts like "is this record explicitly present in
    the input geometry's control points" are separate from concepts like "is
    this record present in the input geometry but not explicitly in its
    control points". The populated fields are also removed, and tracked
    separately in each algorithm that modifies the labels.

## Check List

Have you:

- Added unit tests? No, relies on existing.

- Add cmprefimpl tests? (if appropriate?) No, relies on existing.

- Updated release notes? (if appropriate?) No, relies on existing.

## Related Issue

- N/A

## Benchmark Results

- No significant changes.

<details>

<summary>Click to expand</summary>

```
name                                                        old time/op    new time/op    delta
pkg:github.com/peterstace/simplefeatures/geom goos:linux goarch:arm64
LineEnvelope/0-2                                              2.22ns ±10%    2.25ns ± 8%    ~     (p=0.361 n=15+15)
LineEnvelope/1-2                                              2.17ns ± 6%    2.18ns ± 5%    ~     (p=0.720 n=15+15)
LineEnvelope/2-2                                              2.13ns ± 3%    2.14ns ± 6%    ~     (p=0.710 n=14+13)
LineEnvelope/3-2                                              2.14ns ± 4%    2.14ns ± 6%    ~     (p=0.442 n=15+14)
MarshalWKB/polygon/n=10-2                                     93.3ns ± 4%    92.3ns ± 1%    ~     (p=0.054 n=14+14)
MarshalWKB/polygon/n=100-2                                     237ns ± 4%     237ns ± 6%    ~     (p=0.847 n=15+14)
MarshalWKB/polygon/n=1000-2                                   1.63µs ±19%    1.60µs ±12%    ~     (p=0.949 n=15+14)
MarshalWKB/polygon/n=10000-2                                  40.4µs ±79%    37.7µs ±76%    ~     (p=0.713 n=15+15)
UnmarshalWKB/polygon/n=10-2                                    156ns ± 1%     157ns ± 1%  +0.73%  (p=0.012 n=13+13)
UnmarshalWKB/polygon/n=100-2                                   297ns ± 6%     300ns ± 8%    ~     (p=0.591 n=14+15)
UnmarshalWKB/polygon/n=1000-2                                 1.80µs ± 9%    1.79µs ±10%    ~     (p=0.616 n=15+15)
UnmarshalWKB/polygon/n=10000-2                                42.2µs ±49%    41.0µs ±73%    ~     (p=0.744 n=15+15)
IntersectsLineStringWithLineString/n=10-2                      765ns ± 2%     767ns ± 4%    ~     (p=0.856 n=15+14)
IntersectsLineStringWithLineString/n=100-2                    11.8µs ± 1%    11.9µs ± 4%    ~     (p=0.611 n=14+14)
IntersectsLineStringWithLineString/n=1000-2                    149µs ±15%     153µs ±22%    ~     (p=0.838 n=15+15)
IntersectsLineStringWithLineString/n=10000-2                  3.28ms ±39%    3.36ms ±38%    ~     (p=0.747 n=14+15)
IntersectsMultiPointWithMultiPoint/n=20-2                      555ns ± 1%     553ns ± 1%  -0.47%  (p=0.007 n=14+12)
IntersectsMultiPointWithMultiPoint/n=200-2                    7.54µs ± 1%    7.53µs ± 0%    ~     (p=0.610 n=14+14)
IntersectsMultiPointWithMultiPoint/n=2000-2                   71.2µs ± 1%    71.6µs ± 3%    ~     (p=0.935 n=15+15)
IntersectsMultiPointWithMultiPoint/n=20000-2                   779µs ± 0%     781µs ± 1%    ~     (p=0.949 n=15+14)
PolygonSingleRingValidation/n=10-2                            1.27µs ± 1%    1.27µs ± 1%    ~     (p=0.273 n=13+14)
PolygonSingleRingValidation/n=100-2                           19.1µs ± 1%    19.1µs ± 1%    ~     (p=0.786 n=15+13)
PolygonSingleRingValidation/n=1000-2                           282µs ± 0%     282µs ± 1%    ~     (p=0.621 n=14+15)
PolygonSingleRingValidation/n=10000-2                         4.79ms ±22%    4.60ms ±13%    ~     (p=0.454 n=14+14)
PolygonMultipleRingsValidation/n=4-2                          3.65µs ± 2%    3.67µs ± 3%    ~     (p=0.691 n=15+15)
PolygonMultipleRingsValidation/n=36-2                         30.6µs ± 3%    30.5µs ± 1%    ~     (p=0.335 n=14+14)
PolygonMultipleRingsValidation/n=400-2                         404µs ± 3%     403µs ± 4%    ~     (p=0.325 n=14+13)
PolygonMultipleRingsValidation/n=4096-2                       5.16ms ± 1%    5.17ms ± 1%    ~     (p=0.910 n=14+14)
PolygonZigZagRingsValidation/n=10-2                           6.11µs ± 1%    6.11µs ± 1%    ~     (p=0.725 n=15+13)
PolygonZigZagRingsValidation/n=100-2                          73.6µs ± 1%    73.4µs ± 1%    ~     (p=0.150 n=14+14)
PolygonZigZagRingsValidation/n=1000-2                          983µs ± 5%     984µs ± 5%    ~     (p=0.701 n=14+14)
PolygonZigZagRingsValidation/n=10000-2                        16.7ms ±22%    16.5ms ±24%    ~     (p=0.652 n=15+14)
PolygonAnnulusValidation/n=10-2                               1.97µs ± 1%    1.98µs ± 3%    ~     (p=0.315 n=14+15)
PolygonAnnulusValidation/n=100-2                              19.1µs ± 1%    19.1µs ± 1%    ~     (p=0.801 n=13+13)
PolygonAnnulusValidation/n=1000-2                              340µs ± 2%     340µs ± 2%    ~     (p=0.545 n=13+13)
PolygonAnnulusValidation/n=10000-2                            4.52ms ± 4%    4.53ms ± 7%    ~     (p=0.616 n=14+13)
MultipolygonValidation/n=1-2                                   217ns ± 1%     219ns ± 3%  +1.03%  (p=0.010 n=12+13)
MultipolygonValidation/n=4-2                                   488ns ± 2%     491ns ± 1%  +0.59%  (p=0.048 n=14+14)
MultipolygonValidation/n=16-2                                 2.14µs ± 2%    2.18µs ± 3%  +1.93%  (p=0.000 n=14+15)
MultipolygonValidation/n=64-2                                 10.3µs ± 1%    10.4µs ± 1%  +1.16%  (p=0.000 n=14+12)
MultipolygonValidation/n=256-2                                54.8µs ± 3%    55.9µs ± 3%  +1.92%  (p=0.001 n=14+13)
MultipolygonValidation/n=1024-2                                340µs ± 1%     341µs ± 1%    ~     (p=0.413 n=15+13)
MultiPolygonTwoCircles/n=10-2                                 1.93µs ± 2%    1.93µs ± 2%    ~     (p=0.334 n=14+14)
MultiPolygonTwoCircles/n=100-2                                23.0µs ± 3%    23.1µs ± 2%    ~     (p=0.616 n=13+14)
MultiPolygonTwoCircles/n=1000-2                                326µs ±27%     337µs ±40%    ~     (p=0.856 n=13+15)
MultiPolygonTwoCircles/n=10000-2                              4.81ms ±13%    5.08ms ±15%    ~     (p=0.126 n=15+15)
MultiPolygonMultipleTouchingPoints/n=1-2                      2.60µs ± 1%    2.60µs ± 1%    ~     (p=0.510 n=13+14)
MultiPolygonMultipleTouchingPoints/n=10-2                     20.9µs ± 1%    20.9µs ± 0%    ~     (p=0.205 n=12+13)
MultiPolygonMultipleTouchingPoints/n=100-2                     265µs ± 1%     262µs ± 2%  -1.11%  (p=0.001 n=13+14)
MultiPolygonMultipleTouchingPoints/n=1000-2                   3.29ms ± 3%    3.29ms ± 2%    ~     (p=0.614 n=13+13)
WKTParsing/point-2                                            1.13µs ± 1%    1.13µs ± 1%  -0.46%  (p=0.022 n=15+14)
DistancePolygonToPolygonOrdering/n=100_swap=false-2           26.9µs ± 3%    26.6µs ± 2%    ~     (p=0.256 n=15+14)
DistancePolygonToPolygonOrdering/n=100_swap=true-2            26.5µs ± 1%    26.5µs ± 1%    ~     (p=0.952 n=14+13)
DistancePolygonToPolygonOrdering/n=1000_swap=false-2           462µs ± 1%     463µs ± 1%    ~     (p=0.110 n=13+12)
DistancePolygonToPolygonOrdering/n=1000_swap=true-2            462µs ± 1%     462µs ± 1%    ~     (p=0.561 n=15+14)
IntersectionPolygonWithPolygonOrdering/n=100_swap=false-2     3.26µs ± 4%    3.23µs ± 1%    ~     (p=0.319 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=100_swap=true-2      3.23µs ± 1%    3.23µs ± 1%    ~     (p=0.456 n=13+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=false-2    34.7µs ± 2%    34.6µs ± 1%    ~     (p=0.458 n=14+13)
IntersectionPolygonWithPolygonOrdering/n=1000_swap=true-2     34.6µs ± 1%    35.0µs ± 3%    ~     (p=0.073 n=12+14)
MultiLineStringIsSimpleManyLineStrings/n=100-2                26.7µs ± 2%    26.8µs ± 2%    ~     (p=0.352 n=14+14)
MultiLineStringIsSimpleManyLineStrings/n=1000-2                327µs ± 4%     326µs ± 2%    ~     (p=0.519 n=13+14)
ForceCWandForceCCW/0-2                                        16.7ns ±16%    16.7ns ± 8%    ~     (p=0.573 n=14+13)
ForceCWandForceCCW/0#01-2                                      101ns ± 2%     101ns ± 2%    ~     (p=0.637 n=12+11)
ForceCWandForceCCW/1-2                                        16.8ns ± 8%    16.7ns ± 6%    ~     (p=0.929 n=14+12)
ForceCWandForceCCW/1#01-2                                      101ns ± 2%     101ns ± 1%    ~     (p=0.340 n=13+12)
ForceCWandForceCCW/2-2                                        30.9ns ±10%    31.1ns ± 7%    ~     (p=0.921 n=12+12)
ForceCWandForceCCW/2#01-2                                      155ns ± 1%     155ns ± 1%    ~     (p=0.496 n=9+12)
ForceCWandForceCCW/3-2                                        30.9ns ± 6%    30.5ns ± 9%    ~     (p=0.164 n=13+12)
ForceCWandForceCCW/3#01-2                                      155ns ± 1%     156ns ± 3%    ~     (p=0.096 n=12+10)
ForceCWandForceCCW/4-2                                        45.2ns ± 8%    45.3ns ± 6%    ~     (p=0.524 n=15+12)
ForceCWandForceCCW/4#01-2                                      240ns ± 1%     241ns ± 2%    ~     (p=0.083 n=15+10)
ForceCWandForceCCW/5-2                                        44.2ns ± 7%    44.5ns ± 7%    ~     (p=0.502 n=14+12)
ForceCWandForceCCW/5#01-2                                      241ns ± 1%     240ns ± 0%    ~     (p=0.552 n=14+10)
ForceCWandForceCCW/6-2                                        63.5ns ± 5%    65.5ns ± 8%  +3.25%  (p=0.027 n=13+13)
ForceCWandForceCCW/6#01-2                                      389ns ± 3%     386ns ± 2%    ~     (p=0.135 n=13+13)
EnvelopeTransformXY-2                                          143ns ± 5%     143ns ± 7%    ~     (p=1.000 n=15+15)
WKBParse/0-2                                                   420ns ± 1%     421ns ± 1%    ~     (p=0.480 n=14+13)
WKBParse/1-2                                                   424ns ± 0%     416ns ± 0%  -1.96%  (p=0.000 n=13+13)
WKBParse/2-2                                                   490ns ± 1%     491ns ± 0%    ~     (p=0.303 n=13+13)
WKBParse/3-2                                                   480ns ± 1%     481ns ± 0%  +0.20%  (p=0.010 n=12+14)
WKBParse/4-2                                                   567ns ± 1%     568ns ± 1%    ~     (p=0.626 n=15+13)
WKBParse/5-2                                                   422ns ± 0%     422ns ± 0%  +0.20%  (p=0.011 n=13+13)
WKBParse/6-2                                                   486ns ± 1%     487ns ± 1%    ~     (p=0.079 n=14+14)
WKBParse/7-2                                                   485ns ± 1%     487ns ± 1%  +0.37%  (p=0.038 n=14+14)
WKBParse/8-2                                                   576ns ± 1%     577ns ± 0%    ~     (p=0.113 n=14+13)
WKBParse/9-2                                                   310ns ± 0%     313ns ± 2%  +0.67%  (p=0.001 n=12+13)
WKBParse/10-2                                                  312ns ± 1%     313ns ± 1%    ~     (p=0.243 n=14+13)
WKBParse/11-2                                                  312ns ± 1%     314ns ± 2%  +0.73%  (p=0.001 n=14+14)
WKBParse/12-2                                                  309ns ± 2%     310ns ± 1%  +0.36%  (p=0.036 n=15+14)
WKBParse/13-2                                                  593ns ± 1%     598ns ± 1%  +0.82%  (p=0.000 n=14+14)
WKBParse/14-2                                                  703ns ± 1%     704ns ± 1%    ~     (p=0.105 n=15+13)
WKBParse/15-2                                                  712ns ± 1%     713ns ± 0%    ~     (p=0.190 n=15+14)
WKBParse/16-2                                                  838ns ± 1%     838ns ± 0%    ~     (p=0.765 n=14+13)
WKBParse/17-2                                                  701ns ± 1%     704ns ± 1%  +0.41%  (p=0.011 n=13+15)
WKBParse/18-2                                                  907ns ± 1%     910ns ± 1%  +0.35%  (p=0.002 n=13+14)
WKBParse/19-2                                                  905ns ± 0%     909ns ± 1%  +0.47%  (p=0.000 n=14+14)
WKBParse/20-2                                                 1.10µs ± 0%    1.10µs ± 1%  +0.49%  (p=0.000 n=14+14)
WKBParse/21-2                                                  300ns ± 1%     302ns ± 3%  +0.71%  (p=0.005 n=13+13)
WKBParse/22-2                                                  303ns ± 1%     303ns ± 1%    ~     (p=0.394 n=15+13)
WKBParse/23-2                                                  303ns ± 2%     303ns ± 1%    ~     (p=0.119 n=13+14)
WKBParse/24-2                                                  298ns ± 1%     299ns ± 1%    ~     (p=0.172 n=13+14)
WKBParse/25-2                                                 1.56µs ± 1%    1.56µs ± 1%    ~     (p=0.073 n=14+13)
WKBParse/26-2                                                 1.98µs ± 3%    1.98µs ± 1%    ~     (p=0.473 n=14+14)
WKBParse/27-2                                                 1.97µs ± 1%    1.96µs ± 0%  -0.50%  (p=0.002 n=13+14)
WKBParse/28-2                                                 2.47µs ± 0%    2.47µs ± 1%    ~     (p=0.215 n=12+14)
WKBParse/29-2                                                  301ns ± 1%     302ns ± 0%  +0.44%  (p=0.002 n=15+13)
WKBParse/30-2                                                  303ns ± 2%     305ns ± 1%  +0.70%  (p=0.002 n=15+15)
WKBParse/31-2                                                  302ns ± 1%     304ns ± 1%  +0.72%  (p=0.000 n=15+15)
WKBParse/32-2                                                  298ns ± 0%     300ns ± 1%  +0.82%  (p=0.000 n=13+15)
WKBParse/33-2                                                  569ns ± 0%     569ns ± 1%    ~     (p=0.821 n=14+15)
WKBParse/34-2                                                  649ns ± 1%     649ns ± 1%    ~     (p=0.784 n=13+14)
WKBParse/35-2                                                  658ns ± 1%     649ns ± 1%  -1.38%  (p=0.000 n=13+13)
WKBParse/36-2                                                  710ns ± 0%     712ns ± 1%    ~     (p=0.394 n=14+14)
WKBParse/37-2                                                  814ns ± 1%     815ns ± 0%    ~     (p=0.288 n=13+12)
WKBParse/38-2                                                  958ns ± 2%     960ns ± 1%    ~     (p=0.270 n=14+14)
WKBParse/39-2                                                  966ns ± 2%     954ns ± 1%  -1.19%  (p=0.000 n=14+13)
WKBParse/40-2                                                 1.08µs ± 1%    1.08µs ± 1%    ~     (p=0.808 n=12+12)
WKBParse/41-2                                                  302ns ± 1%     303ns ± 0%    ~     (p=0.057 n=13+12)
WKBParse/42-2                                                  305ns ± 0%     305ns ± 1%    ~     (p=0.838 n=12+15)
WKBParse/43-2                                                  305ns ± 1%     305ns ± 1%    ~     (p=0.890 n=12+14)
WKBParse/44-2                                                  302ns ± 1%     301ns ± 1%    ~     (p=0.510 n=13+14)
WKBParse/45-2                                                  887ns ± 1%     888ns ± 1%    ~     (p=0.379 n=15+12)
WKBParse/46-2                                                 1.07µs ± 1%    1.08µs ± 3%  +0.88%  (p=0.048 n=14+14)
WKBParse/47-2                                                 1.08µs ± 1%    1.08µs ± 1%    ~     (p=0.610 n=14+14)
WKBParse/48-2                                                 1.24µs ± 2%    1.24µs ± 2%    ~     (p=0.541 n=14+13)
WKBParse/49-2                                                 1.26µs ± 1%    1.27µs ± 1%  +0.39%  (p=0.004 n=15+15)
WKBParse/50-2                                                 1.60µs ± 1%    1.60µs ± 1%    ~     (p=0.197 n=14+13)
WKBParse/51-2                                                 1.60µs ± 1%    1.60µs ± 1%    ~     (p=0.407 n=13+14)
WKBParse/52-2                                                 1.88µs ± 1%    1.88µs ± 1%    ~     (p=0.610 n=14+14)
WKBParse/53-2                                                  303ns ± 3%     303ns ± 1%    ~     (p=0.091 n=15+13)
WKBParse/54-2                                                  305ns ± 1%     305ns ± 1%    ~     (p=0.533 n=14+14)
WKBParse/55-2                                                  304ns ± 2%     304ns ± 1%    ~     (p=0.563 n=13+15)
WKBParse/56-2                                                  299ns ± 1%     300ns ± 1%    ~     (p=0.100 n=12+15)
WKBParse/57-2                                                 1.91µs ± 1%    1.92µs ± 1%  +0.38%  (p=0.017 n=14+13)
WKBParse/58-2                                                 2.36µs ± 1%    2.37µs ± 0%  +0.34%  (p=0.001 n=14+13)
WKBParse/59-2                                                 2.35µs ± 0%    2.36µs ± 1%  +0.53%  (p=0.000 n=13+13)
WKBParse/60-2                                                 2.85µs ± 1%    2.87µs ± 1%  +0.46%  (p=0.037 n=15+13)
WKBParse/61-2                                                  302ns ± 0%     302ns ± 1%    ~     (p=0.772 n=13+13)
WKBParse/62-2                                                  304ns ± 1%     304ns ± 1%    ~     (p=0.574 n=13+14)
WKBParse/63-2                                                  304ns ± 1%     304ns ± 1%    ~     (p=0.525 n=14+15)
WKBParse/64-2                                                  300ns ± 2%     299ns ± 0%    ~     (p=0.640 n=13+13)
WKBParse/65-2                                                  871ns ± 5%     863ns ± 1%    ~     (p=0.627 n=14+14)
WKBParse/66-2                                                 1.02µs ± 2%    1.02µs ± 1%    ~     (p=0.758 n=13+15)
WKBParse/67-2                                                 1.02µs ± 2%    1.02µs ± 1%    ~     (p=0.721 n=14+15)
WKBParse/68-2                                                 1.13µs ± 2%    1.12µs ± 1%  -1.01%  (p=0.000 n=14+13)
pkg:github.com/peterstace/simplefeatures/geos goos:linux goarch:arm64
IntersectionWithoutValidation/n=10-2                          13.4µs ± 2%    13.4µs ± 2%    ~     (p=0.683 n=15+14)
IntersectionWithoutValidation/n=100-2                         29.0µs ± 2%    29.0µs ± 2%    ~     (p=0.880 n=15+14)
IntersectionWithoutValidation/n=1000-2                         148µs ± 2%     149µs ± 4%    ~     (p=0.126 n=15+15)
…