@peterstace 2022-11-11
> At a high level, from what I understand, the aim of this PR is to improve how we model the state of a record by tailoring the state data (labels) to be more specific to the record type (vertex, half edge or face). Is that correct?

Yes, that's right. The original implementation reused the labels struct or each of the 3 record types which saved a bit of code but was a fairly inaccurate/confusing modelling.